conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aiohttp=3.8.5=pypi_0
aiosignal=1.3.1=pypi_0
ansicon=1.89.0=pypi_0
anyio=3.7.1=pypi_0
appdirs=1.4.4=pypi_0
arrow=1.2.3=pypi_0
async-timeout=4.0.2=pypi_0
attrs=23.1.0=pypi_0
beautifulsoup4=4.12.2=pypi_0
blas=1.0=mkl
blessed=1.20.0=pypi_0
brotlipy=0.7.0=py39h2bbff1b_1003
ca-certificates=2023.05.30=haa95532_0
certifi=2023.7.22=py39haa95532_0
cffi=1.15.1=py39h2bbff1b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.6=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.1.0=pypi_0
croniter=1.3.15=pypi_0
cryptography=41.0.2=py39hac1b9e3_0
cuda-cccl=12.2.128=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cycler=0.11.0=pypi_0
dateutils=0.6.12=pypi_0
deepdiff=6.3.1=pypi_0
docker-pycreds=0.4.0=pypi_0
exceptiongroup=1.1.2=pypi_0
fastapi=0.88.0=pypi_0
filelock=3.9.0=py39haa95532_0
fonttools=4.41.1=pypi_0
freetype=2.12.1=ha860e81_0
frozenlist=1.4.0=pypi_0
fsspec=2023.6.0=pypi_0
giflib=5.2.1=h8cc25b3_3
gitdb=4.0.10=pypi_0
gitpython=3.1.32=pypi_0
h11=0.14.0=pypi_0
idna=3.4=py39haa95532_0
importlib-resources=6.0.0=pypi_0
inquirer=3.1.3=pypi_0
intel-openmp=2023.1.0=h59b6b97_46319
itsdangerous=2.1.2=pypi_0
jinja2=3.1.2=py39haa95532_0
jinxed=1.2.0=pypi_0
joblib=1.3.1=pypi_0
jpeg=9e=h2bbff1b_1
kiwisolver=1.4.4=pypi_0
lame=3.100=hcfcfb64_1003
lerc=3.0=hd77b12b_0
libcublas=11.10.3.66=0
libcublas-dev=11.10.3.66=0
libcufft=10.7.2.124=0
libcufft-dev=10.7.2.124=0
libcurand=10.3.3.129=0
libcurand-dev=10.3.3.129=0
libcusolver=11.4.0.1=0
libcusolver-dev=11.4.0.1=0
libcusparse=11.7.4.91=0
libcusparse-dev=11.7.4.91=0
libdeflate=1.17=h2bbff1b_0
libflac=1.4.3=h63175ca_0
libnpp=11.7.4.75=0
libnpp-dev=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libnvjpeg-dev=11.8.0.2=0
libogg=1.3.4=h8ffe710_1
libopus=1.3.1=h8ffe710_1
libpng=1.6.39=h8cc25b3_0
libsndfile=1.2.0=h2628c91_0
libtiff=4.5.0=h6c2663c_2
libuv=1.44.2=h2bbff1b_0
libvorbis=1.3.7=h0e60522_0
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
lightning=1.9.5=pypi_0
lightning-cloud=0.5.37=pypi_0
lightning-utilities=0.9.0=pypi_0
lz4-c=1.9.4=h2bbff1b_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.1=py39h2bbff1b_0
matplotlib=3.7.2=pypi_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h8bd8f75_46356
mkl-service=2.4.0=py39h2bbff1b_1
mkl_fft=1.3.6=py39hf11a4ad_1
mkl_random=1.2.2=py39hf11a4ad_1
mpg123=1.31.3=h63175ca_0
mpmath=1.3.0=py39haa95532_0
multidict=6.0.4=pypi_0
networkx=3.1=py39haa95532_0
numpy=1.25.0=py39h055cbcc_0
numpy-base=1.25.0=py39h65a83cf_0
opencv-python=4.8.0.74=pypi_0
openssl=3.0.10=h2bbff1b_2
ordered-set=4.1.0=pypi_0
packaging=23.1=pypi_0
pandas=2.0.3=pypi_0
pathtools=0.1.2=pypi_0
pillow=9.4.0=py39hd77b12b_0
pip=23.2.1=py39haa95532_0
protobuf=4.23.4=pypi_0
psutil=5.9.5=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pydantic=1.10.12=pypi_0
pygments=2.15.1=pypi_0
pyjwt=2.8.0=pypi_0
pyopenssl=23.2.0=py39haa95532_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py39haa95532_0
pysoundfile=0.12.1=pyhd8ed1ab_0
python=3.9.17=h1aa4202_0
python-dateutil=2.8.2=pypi_0
python-editor=1.0.4=pypi_0
python-multipart=0.0.6=pypi_0
pytorch=2.0.1=py3.9_cuda11.7_cudnn8_0
pytorch-cuda=11.7=h16d0643_5
pytorch-lightning=1.9.5=pypi_0
pytorch-mutex=1.0=cuda
pytz=2023.3=pypi_0
pyyaml=6.0.1=pypi_0
readchar=4.0.5=pypi_0
requests=2.31.0=py39haa95532_0
rich=13.5.2=pypi_0
scikit-learn=1.3.0=pypi_0
scipy=1.11.1=pypi_0
seaborn=0.12.2=pypi_0
sentry-sdk=1.29.2=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=68.0.0=py39haa95532_0
six=1.16.0=pypi_0
sklearn=0.0.post7=pypi_0
smmap=5.0.0=pypi_0
sniffio=1.3.0=pypi_0
soupsieve=2.4.1=pypi_0
sqlite=3.41.2=h2bbff1b_0
starlette=0.22.0=pypi_0
starsessions=1.3.0=pypi_0
sympy=1.11.1=py39haa95532_0
tbb=2021.8.0=h59b6b97_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.12=h2bbff1b_0
torchaudio=2.0.2=pypi_0
torchinfo=1.8.0=pypi_0
torchmetrics=1.0.1=pypi_0
torchvision=0.15.2=pypi_0
tqdm=4.65.0=pypi_0
traitlets=5.9.0=pypi_0
typing_extensions=4.7.1=py39haa95532_0
tzdata=2023.3=pypi_0
ucrt=10.0.22621.0=h57928b3_0
urllib3=1.26.16=py39haa95532_0
uvicorn=0.23.2=pypi_0
vc=14.2=h21ff451_1
vc14_runtime=14.36.32532=hfdfe4a8_17
vs2015_runtime=14.36.32532=h05e6639_17
wandb=0.15.8=pypi_0
wcwidth=0.2.6=pypi_0
websocket-client=1.6.1=pypi_0
websockets=11.0.3=pypi_0
wheel=0.38.4=py39haa95532_0
win_inet_pton=1.1.0=py39haa95532_0
xz=5.4.2=h8cc25b3_0
yarl=1.9.2=pypi_0
zipp=3.16.2=pypi_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
